chore: replace `flights.py`
- deletes `flights.py`
- renames `flights2.py` -> `flights.py`
